Create README.md
So I wanted to try to code something that will work for my final project I intend to do, which is lights on a bike so people can see you at night. With all the work with blinking leds that we were doing I had an Idea, why not code for one light to blink in a certain state and to stay solid in another. A lot of the code was borrowed from examples but I did face a challenge with figuring out how to get 2 lights working with the remixed code of a button switching between a solid and a blinking led. I chose red to blink because red means stop and the blinking is more apparent then a solid. The solid is for just to be visible during the night.